38/68 Catching up with my commits

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -1,10 +1,34 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
+const { randomBytes } = require('crypto');
+const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // TODO: Check that they are logged in.
-    const item = await ctx.db.mutation.createItem({ data: { ...args } }, info);
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+    console.log(ctx.request.userId, args);
+
+    const item = await ctx.db.mutation.createItem(
+      {
+        data: {
+          // This is how to create a relationship between the Item and the User
+          user: {
+            connect: {
+              id: ctx.request.userId
+            }
+          },
+          ...args
+        }
+      },
+      info
+    );
+
+    console.log(item);
+
     return item;
   },
 
@@ -20,9 +44,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. Find item
-    const item = await ctx.db.query.item({ where }, '{id title}');
+    const item = await ctx.db.query.item({ where }, '{id title user { id }}');
     // 2. Check if the user owns the item, or have permissions
-
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
     // 3. Delete Item
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -30,7 +60,7 @@ const Mutations = {
   async signup(parent, args, ctx, info) {
     // To make all signup usages the same, lowercase the email
     args.email = args.email.toLowerCase();
-    // has the password
+    // hash the password
     const password = await bcrypt.hash(args.password, 10);
     // create user in db
     const user = await ctx.db.mutation.createUser(
@@ -80,6 +110,89 @@ const Mutations = {
   signout(parent, args, ctx, info) {
     ctx.response.clearCookie('token');
     return { message: 'Goodbye!' };
+  },
+  async requestReset(parent, args, ctx, info) {
+    // 1. Check if it is a real user
+    const user = await ctx.db.query.user({ where: { email: args.email } });
+    if (!user) {
+      throw new Error(`No such user found for email ${user.email}`);
+    }
+    // 2. Set a reset token and expiry on that user
+    const randomBytesPromisified = promisify(randomBytes);
+    const resetToken = (await randomBytesPromisified(20)).toString('hex');
+    const resetTokenExpiry = Date.now() + 3600000; // 1 hour from now;
+    await ctx.db.mutation.updateUser({
+      where: { email: args.email },
+      data: { resetToken, resetTokenExpiry }
+    });
+    // 3. Email them that reset token
+    const mailResponse = await transport.sendMail({
+      from: 'cjared.mortenson@gmail.com',
+      to: user.email,
+      subject: 'Your password reset token',
+      html: makeANiceEmail(
+        `Your Password Reset is Here! \n\n
+        <a href="${process.env.FRONTEND_URL}/reset?resetToken=${resetToken}">
+          Click here to reset your password!
+        </a>`
+      )
+    });
+
+    return { message: 'You need to check your email now.' };
+  },
+  async resetPassword(parent, args, ctx, info) {
+    // 1. Check if passwords match
+    if (args.password !== args.confirmPassword) {
+      throw new Error("Your passwords don't match!");
+    }
+    // 2. Check if legit reset token
+    // 3. check if it is expeired
+    const [user] = await ctx.db.query.users({
+      where: {
+        resetToken: args.resetToken,
+        resetTokenExpiry_gte: Date.now() - 3600000
+      }
+    });
+    if (!user) {
+      throw new Error('This token is either invalid or expired!');
+    }
+    // 4. hash their new password
+    const password = await bcrypt.hash(args.password, 10);
+    // 5. save the new password to the user and remove old resetToken fields
+    const updatedUser = await ctx.db.mutation.updateUser({
+      where: { email: user.email },
+      data: { password, resetToken: null, resetTokenExpiry: null }
+    });
+    // 6. Generate JWT
+    const token = jwt.sign({ userId: updatedUser.id }, process.env.APP_SECRET);
+    // 7. set the JWT cookie
+    ctx.response.cookie('token', token, {
+      httpOnly: true,
+      maxAge: 1000 * 60 * 60 * 24 * 365 // One year cookie
+    });
+    // 8. Return the new user
+    return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+    // 2. Query the current user
+    const currentUser = await ctx.db.query.user(
+      { where: { id: ctx.request.userId } },
+      info
+    );
+    // 3. Check if they have permission to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. Update
+    return ctx.db.mutation.updateUser(
+      {
+        data: { permissions: { set: args.permissions } },
+        where: { id: args.userId }
+      },
+      info
+    );
   }
 };
